spark/reconciler_agent/pyproject.toml
[project]
name = "reconciler-agent"
version = "0.1.0"
description = "Reconciler Agent for automated transaction discrepancy resolution"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=0.2.0",
    "google-adk>=0.1.0",
    "a2a-sdk>=0.1.0",
    "python-dotenv>=1.0.0",
    "uvicorn>=0.23.2",
    "starlette>=0.27.0",
    "pydantic>=2.0.0",
    "asyncpg>=0.28.0",
    "httpx>=0.24.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []